+++ app/src/main/java/com/enriquebecerra/snaketracker/SnakeTrackerApplication.kt
@@ -3,6 +3,7 @@ package com.enriquebecerra.snaketracker
 import android.app.Application
 import com.enriquebecerra.snaketracker.data.local.database.SnakeTrackerDatabase
 import com.enriquebecerra.snaketracker.data.repository.FeedingRepository
+import com.enriquebecerra.snaketracker.data.repository.LengthRepository
 import com.enriquebecerra.snaketracker.data.repository.PetRepository
 import com.enriquebecerra.snaketracker.data.repository.WeightRepository
 
@@ -22,5 +23,9 @@ class SnakeTrackerApplication : Application() {
 
     val weightRepository: WeightRepository by lazy {
         WeightRepository(database.weightLogDao())
+    }
+
+    val lengthRepository: LengthRepository by lazy {
+        LengthRepository(database.lengthLogDao())
     }
 }